requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
colorama==0.4.6
fastapi==0.104.1
h11==0.14.0
idna==3.10
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
sqlalchemy==2.0.27
pymysql==1.1.0
